added course selection page
